Split test_app into smaller files.
